Staging

+++ src/Controller/UserController.php
@@ -45,11 +45,12 @@ class UserController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-
             $a = $fileUpload->upload($form->get('picture')->getData(), 'photo de profil de ' . $user->getPseudo(), 'file', 'image', 'setUserImage', $user);
 
             $user->getPicture();
-            $entityManager->remove( $user->getPicture());
+            if ($user->getPicture()) {
+                $entityManager->remove($user->getPicture());
+            }
             $entityManager->flush();
             $user->setPicture($a);
             $entityManager->persist($user);
